requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.2
django-chartjs==2.3.0
django-htmx==1.17.3
django-humanize==0.1.2
django-jalali==7.1.0
django-widget-tweaks==1.5.0
humanize==4.9.0
jalali_core==1.0.0
jdatetime==5.0.0
sqlparse==0.5.0
